+++ unaos/crates/kernel/src/arch/x86_64/syscall.rs
@@ -23942,17 +23942,17 @@ fn owned_user_ok(nameid: usize, slot: usize) -> bool {
 /// SO37: the LIVE session's epoch (starts at 1; 0 is the never-stamped slot value). Bumped by every
 /// [`session_logout`].
 #[cfg(feature = "login")]
-static SESSION_EPOCH: AtomicU32 = AtomicU32::new(1);
+static SESSION_EPOCH: AtomicU64 = AtomicU64::new(1); // SECLOGIN M5: u64 — a u32 wraps at 4.3e9 logouts, named by B157 gap 5, now defended
 
 /// SO37: the epoch each slot's user stamp was taken in (0 = never stamped, which never matches).
 #[cfg(feature = "login")]
-static SLOT_EPOCH: [AtomicU32; crate::arch::memory::USER_SLOTS + 1] =
-    [const { AtomicU32::new(0) }; crate::arch::memory::USER_SLOTS + 1];
+static SLOT_EPOCH: [AtomicU64; crate::arch::memory::USER_SLOTS + 1] =
+    [const { AtomicU64::new(0) }; crate::arch::memory::USER_SLOTS + 1]; // SECLOGIN M5: u64 with SESSION_EPOCH
 
 /// SO37: the LIVE epoch as a NUMBER, for the wire — never a decision. Read by `fs::users::logout`, so
 /// every Log Out names the epoch it just opened and a boot's session boundaries are countable on serial.
 #[cfg(feature = "login")]
-pub fn session_epoch() -> u32 {
+pub fn session_epoch() -> u64 {
     SESSION_EPOCH.load(Ordering::Acquire)
 }
 
@@ -25295,7 +25295,7 @@ pub fn ident_fixture(uid_a: u32, uid_b: u32, uid_a2: u32) -> (bool, bool, bool) 
 /// the boundary is the session, and an anonymous program has no session to outlive. Returns
 /// `(ended, windows_closed)`; called from [`session_logout`] BEFORE the epoch bump.
 #[cfg(feature = "login")]
-fn session_end_processes(closing: u32) -> (usize, usize) {
+fn session_end_processes(closing: u64) -> (usize, usize) {
     let mut ended = 0usize;
     let mut windows = 0usize;
     for pi in 0..MAX_PROCS {
